7-average_score.sql

+++ 0x00-MySQL_Advanced/7-average_score.sql
@@ -6,7 +6,7 @@ DELIMITER //
 
 CREATE PROCEDURE ComputeAverageScoreForUser(IN user_id INT)
 BEGIN
-    DECLARE avg_score FLOAT;
+    DECLARE avg_score DECIMAL(10, 2);
 
     -- Calculate the average score for the given user_id
     SELECT AVG(score) INTO avg_score
